% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ecoliepidata}
\alias{Ecoliepidata}
\title{Escherichia coli O157:H7 multi-strain longitudinal data in cattle.}
\format{
A dataframe with 4266 rows (one for each individual per sampling
time) and 5 variables:
\describe{
\item{Time}{An integer representing the time index of each sample,
starting from one for each individual with maximum being 99 (the last
observation time of the study).}
\item{Individual}{An integer representing the animal index, ranging from 1
to 8 for each pen.}
\item{Group}{An integer representing the pen index, ranging from 1 to 20.}
\item{RAJ_PFGE}{Observed result of RAJ test, "+" or "-" or an integer
ranging from 1 to 8.}
\item{Fecal_PFGE}{Observed result of fecal test, "+" or "-" or an
integer ranging from 1 to 8.}
}
}
\usage{
data(Ecoliepidata)
}
\description{
Multi-strain epidemic data obtained from a longitudinal study of
\emph{Escherichia coli} O157:H7 in feedlot cattle in which
7 main competing strains/genotypes were identified. In this study, 160
cattle were randomly allocated to 20 experimental research pens (8 animals
each) located at Canada and were sampled approximately twice per week over
a 14-week period. Briefly, at each sampling time point two samples were
collected from each animal: a fecal sample and rectoanal mucosal swabs
(RAMS). The presence or absence of \emph{E. coli} O157:H7 in each sample
was determined. In addition, some positive samples were randomly selected
from each pen to be genotyped using Pulsed Field Gel Electrophoresis
(PFGE) method.

Therefore, the observed longitudinal data comprises of a set of result
sequences from the two different tests, namely rectoanal junction (RAJ)
and fecal test. More specifically, the test result of each individual was
recorded at each sampling time as "-", if the result was negative, "+"
if it was positive but not chosen for genotyping, otherwise, the genotype
number is given ranging from 1 to 8 (1-7 refers to carriage of one of the
main genotypes and 8 to carriage of the remaining genotypes). For the full
details see references.
}
\details{
Overall, the dataset consists of 4266 pairs of faecal and RAMS samples
derived from sampling 160 cattle on a maximum of 27 different sampling
times throughout the entire study period. All animals remained in their
assigned pens until the end of the study, except four cattle that were
withdrawn due to physical or behavioral problems in repeated handling.
Moreover, animals were housed in North and South pens measuring
6m\eqn{\times}{x}17m and 6m\eqn{\times}{x}37m, respectively. Full details
of the design and conditions of the experiment, as well as the collected
data are described in details in \insertCite{Touloupou2020;textual}{epiPOMS}.
}
\examples{
# Load E. coli O157:H7 data
data(Ecoliepidata)

# Format the data
epidata <- as_epiPOMSdata(Ecoliepidata)

# Plot the data
plot(epidata)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{plot.epiPOMSdata}} for plotting an object of class
produced by \code{\link{as_epiPOMSdata}} function and
\code{\link{epiPOMS_mcmc}} for performing inference in partially observed
multi-strain epidemic models.
}
\author{
Panayiota Touloupou
}
\keyword{datasets}
