% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidstate_sim.R
\name{hidstate_sim}
\alias{hidstate_sim}
\title{Simulate the hidden process (the true carriage states).}
\usage{
hidstate_sim(ntypes, ngroups, ninds, tmax, hidpars, smallgroups, indNA, tNA)
}
\arguments{
\item{ntypes}{An integer greater than 1 providing the number of different strains in
the study, \eqn{n_g}.}

\item{ngroups}{A positive integer providing the total number of groups in the study,
\eqn{P}.}

\item{ninds}{A vector of integers greater than 1 with length \eqn{P}, where
element \eqn{p} gives the number of individuals in group \eqn{p}.}

\item{tmax}{A vector of integers greater than 1 with length \eqn{P}, where
element \eqn{p} gives the time point that the last sample was collected from
individuals in group \eqn{p}.}

\item{hidpars}{A list of parameters with which the simulations of the hidden
process are to be performed. The names of the components must be "alpha",
"beta", "mu", "delta", "gamma" and "nu" (see example). Also, all rates must
be specified in days. These components are:
\describe{
\item{alpha}{A vector of non-negative real numbers, one for each strain,
containing the \eqn{\alpha} parameters, i.e. the strain-specific
external colonisation rates. This is a vector of \eqn{n_g} values.
The first element corresponds to the first strain, the second element to
the second strain and so on.}
\item{beta}{A vector of non-negative real numbers, one for each strain,
containing the \eqn{\beta} parameters, i.e. the strain-specific
within-group colonisation rates. This is a vector of \eqn{n_g} values.
The first element corresponds to the first strain, the second element to
the second strain and so on.}
\item{mu}{A vector of non-negative real numbers, one for each strain,
containing the \eqn{\mu} parameters, i.e. the strain-specific clearance
rates. This is a vector of \eqn{n_g} values.
The first element corresponds to the first strain, the second element to
the second strain and so on.}
\item{delta}{A non-negative real number corresponding to the \eqn{\delta}
parameter, i.e. the relative colonisation rate in a carrier versus
non-carrier individual.}
\item{gamma}{A non-negative real number corresponding to the \eqn{\gamma}
parameter, i.e. the relative colonisation rate in smaller versus bigger
groups in terms of area (in square meters). If
\code{smallgroups = NULL}, i.e there is no difference between groups, set
\code{gamma = NULL}.}
\item{nu}{A vector of numbers between 0 and 1, containing the \eqn{\nu}
parameters, i.e. the probabilities of carriage at the beginning of the study.
This is a vector of \eqn{(n_g + 1)} values. The first element corresponds to
the non-carriage state. The second and subsequent elements correspond to the
carriage of one of the \eqn{n_g} strains. These probabilities should sum up to
one.}
}}

\item{smallgroups}{A vector of positive integers, with length equal to the number
of groups that are smaller in terms of area (in square meters), containing
the indices of the small groups. These values must be between 1 and \eqn{P}.
Set \code{smallgroups = NULL} if there is no difference between groups.}

\item{indNA}{A vector of non-negative integers with length \eqn{P}, where element
\eqn{p} gives the index of the individual that withdrawn from the study
before its completion in group \eqn{p}, if any. Otherwise a value
of 0 should be given. Note that at most one individual from each group can
drop-out before the study period ends.}

\item{tNA}{A vector of positive integers with length \eqn{P}, where element \eqn{p}
provides the individual's drop-out time point in group \eqn{p}, if any,
otherwise the time of the last observation in group \eqn{p} should be given,
\code{tmax}\eqn{[p]}. Note that, if a value is less than the time
of the last observation the corresponding value in \code{indNA} must be
greater than 0, i.e. if \code{tNA}\eqn{[p]} < \code{tmax}\eqn{[p]}
then \code{indNA}\eqn{[p]} > 0.}
}
\value{
A list containing \eqn{P} matrices (one for each group), with
\code{ninds}\eqn{[p]} rows and \code{tmax}\eqn{[p]} columns
for each \eqn{p = 1, 2, \ldots, P}. More specifically, if \eqn{h} is the
returned list, \eqn{h[[p]]} is the matrix of group \eqn{p} containing the
simulated carriage states (of the unobserved process) of each individual in
the group at each time point over the sampling period, where the time interval
is equal to one day. Therefore, \eqn{h[[p]][c, t]} provides the carriage state
of individual \eqn{c = 1, 2, \ldots,} \code{ninds}\eqn{[p]} at time
\eqn{t = 1, 2, \ldots,} \code{tmax}\eqn{[p]}, with \code{NA} entries
corresponding to missing values, for example, if the individual withdrawn from
the study before its completion.
}
\description{
This function simulates the (true) hidden carriage states of each
individual per group during its sampling period.
}
\details{
Our approach to analyse infection dynamics in longitudinal studies where
the process is only partially observed (in the sense that the data are
subject to potential testing error due to poor sensitivity of the
diagnostic and strain procedure used) involves assuming that the
classifications at an observation time are imperfect measures of an underlying
true (hidden) epidemic process. Moreover, our approach accounts for missing
observations, for example, sparse sampling intervals or individual dropouts.

The \code{\link{hidstate_sim}} function simulates the true unobserved
(hidden) disease process within each group (the carriage process is
assumed to be independent across groups), which is modelled as a multi-state,
discrete time non homogeneous Markov model. The possible states include being
a non-carrier (state 0) or being a carrier of one of the \eqn{n_g}
strains (states \eqn{1, 2, \ldots, n_g}), where 1 to
\eqn{(n_g - 1)} refers to carriage of one of the main strains
and \eqn{n_g} to carriage of the remaining strains (pooled group).
A detailed description of this model is found in
\insertCite{Touloupou2020;textual}{epiPOMS}.
}
\examples{
# Simulate the hidden carriage state process of individuals housed in
# 15 different groups of size varied from 4 to 9. The possible states
# include being a non-carrier (state 0) or being a carrier of one of
# 5 strains (states 1, 2, 3, 4 and 5).
set.seed(1)
ntypes <- 5
ngroups <- 15
ninds <- sample(4:9, size = ngroups, replace = TRUE)
tmax <- rep(100, ngroups)
hidpars <- list(alpha = c(sample(c(0.0015, 0.001), ntypes - 1, replace = TRUE),
     0.002), beta = sample(c(0.005, 0.01), ntypes, replace = TRUE), mu =
     sample(c(0.1, 0.15), ntypes, replace = TRUE), delta = 0.5, gamma = 2,
     nu = c(0.9, rep((1-0.9)/(ntypes+1), (ntypes-1)), 2*(1-0.9)/(ntypes+1)))
smallgroups <- sample(1:ngroups, size = 10)
indNA <- c(rep(0, ngroups-1), 1)
tNA <- c(rep(100, ngroups-1), 70)
hiddenprocess <- hidstate_sim(ntypes = ntypes, ngroups = ngroups,
     ninds = ninds, tmax = tmax, hidpars = hidpars, smallgroups =
     smallgroups, indNA = indNA, tNA = tNA)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{obserdata_sim}} for simulating partially observed
multi-strain epidemic data.
}
\author{
Panayiota Touloupou, Simon Spencer
}
