% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMC_upd.R
\name{HMC_upd}
\alias{HMC_upd}
\title{Update the transmission parameters using an HMC algorithm.}
\usage{
HMC_upd(ntypes, ngroups, ninds, tmax, logcuralpha, logcurbeta, logcurmu,
  logcurdelta, logcurgamma, curhidproc, stepsize, nsteps, mass_diag,
  priorcontrol, smallgroups)
}
\arguments{
\item{ntypes}{An integer greater than 1 providing the number of different strains in
the study, \eqn{n_g}.}

\item{ngroups}{A positive integer providing the total number of groups in the study,
\eqn{P}.}

\item{ninds}{A vector of integers greater than 1 with length \eqn{P}, where
element \eqn{p} gives the number of individuals in group \eqn{p}.}

\item{tmax}{A vector of integers greater than 1 with length \eqn{P}, where
element \eqn{p} gives the time point that the last sample was collected from
individuals in group \eqn{p}.}

\item{logcuralpha}{A vector of real numbers, one for each strain, containing the
current values of the logarithm of \eqn{\alpha} parameters, i.e. the
strain-specific external colonisation rates. This is a vector of \eqn{n_g}
values. The first element corresponds to the first strain, the second element
to the second strain and so on.}

\item{logcurbeta}{A vector of real numbers, one for each strain, containing the
current values of the logarithm of \eqn{\beta} parameters, i.e. the
strain-specific within-group colonisation rates. This is a vector of \eqn{n_g}
values. The first element corresponds to the first strain, the second element
to the second strain and so on.}

\item{logcurmu}{A vector of real numbers, one for each strain, containing the
current values of the logarithm of \eqn{\mu} parameters, i.e. the
strain-specific clearance rates. This is a vector of \eqn{n_g} values.
The first element corresponds to the first strain, the second element to
the second strain and so on.}

\item{logcurdelta}{A real number corresponding to the current value of the logarithm
of \eqn{\delta} parameter, i.e. the relative colonisation rate in a
carrier versus non-carrier individual.}

\item{logcurgamma}{A real number corresponding to the current value of the logarithm
of \eqn{\gamma} parameter, i.e. the relative colonisation rate in
smaller versus bigger groups in terms of area (in square meters). If
\code{smallgroups = NULL}, i.e there is no difference between groups, set
\code{logcurgamma = NULL}.}

\item{curhidproc}{A list containing \eqn{P} matrices (one for each group), with
\code{ninds}\eqn{[p]} rows and \code{tmax}\eqn{[p]} columns
for each \eqn{p = 1, 2, \ldots, P}. Each matrix contains the current carriage
states of all individuals within the group at each time point over their
group observation period. All elements must be non-negative integers, ranging
from 0 to \eqn{n_g}, or \code{NA}. \code{NA} entries corresponds to missing
values, for example, individual dropouts.}

\item{stepsize}{A positive real value specifying the leapfrog stepsize in HMC
method.}

\item{nsteps}{A positive integer corresponding to the number of leapfrog steps in
HMC method.}

\item{mass_diag}{A vector of positive numbers providing the diagonal of the mass
matrix, with length equal to the number of parameters to be updated jointly
using the HMC algorithm, i.e. \eqn{(3 \times n_g + 2)}{(3xn_g + 2)} if
\code{smallgroups} \eqn{\neq}{≠} \code{NULL} or
\eqn{(3 \times n_g + 1)}{(3xn_g + 1)} if \code{smallgroups = NULL}. See
details, below, for more information.}

\item{priorcontrol}{A list containing the hyperparameters for the prior
distributions for the MCMC algorithm, obtained using function
\code{\link{prior_control}}.}

\item{smallgroups}{A vector of positive integers, with length equal to the number
of groups that are smaller in terms of area (in square meters), containing
the indices of the small groups. These values must be between 1 and \eqn{P}.
Set \code{smallgroups = NULL} if there is no difference between groups.}
}
\value{
A list is returned with the HMC output of the following updated model
parameters:
\describe{
\item{logcuralpha}{A vector with the updated values of the logarithm of
\eqn{\alpha} parameters.}
\item{logcurbeta}{A vector with the updated values of the logarithm of
\eqn{\beta} parameters.}
\item{logcurmu}{A vector with the updated values of the logarithm of
\eqn{\mu} parameters.}
\item{logcurdelta}{The updated value of the logarithm of
\eqn{\delta} parameter.}
\item{logcurgamma}{The updated value of the logarithm of
\eqn{\gamma} parameter, if \code{smallgroups} \eqn{\neq}{≠} \code{NULL}.
Otherwise a \code{NULL} value is returned.}
}
}
\description{
This function updates the transmission parameters using an Hamiltonian Monte
Carlo (HMC) algorithm. This function is only used in conjunction with
\code{\link{epiPOMS_mcmc}}.
}
\details{
Sampling from the posterior distribution of the parameters of partially
observed multi-strain epidemic models is performed within a Bayesian
framework using an MCMC algorithm (see
\insertCite{Touloupou2020;textual}{epiPOMS} for more details). Some of the
full conditionals are not given in closed form and therefore for these we
resort to HMC \insertCite{Neal2011}{epiPOMS}. More specifically, the following
parameters are updated jointly, in order, with HMC: the strain-specific
external colonisation rates (\eqn{\alpha}), the within-group colonisation
rates (\eqn{\beta}), the clearance rates (\eqn{\mu}), the relative
colonisation rate in a carrier versus non-carrier individual (\eqn{\delta})
and, if \code{smallgroups} \eqn{\neq}{≠} \code{NULL}, the relative colonisation
rate in smaller versus bigger groups in terms of area (\eqn{\gamma}). An
important constraint on these model parameters is that they must be positive
quantities. Additionally, some of the rates may become very small and therefore
we follow the common practice of using a logarithmic transformation for these
parameters.

In the sampling process of HMC, leapfrog integration is used to
approximately solve Hamilton's equations, and the integration is required
to set the number of discrete time steps (\code{nsteps}) and the integration
stepsize (\code{stepsize}). Moreover, the specification of a mass matrix is
required which should be symmetric and positive-definite, and is often a scalar
multiple of the identity matrix. Here, we assume that the mass matrix is
diagonal, with diagonal elements being the elements of \code{mass_diag}.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{epiPOMS_mcmc}} for performing inference in partially observed
multi-strain epidemic  models.
}
\author{
Panayiota Touloupou
}
