% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_epiPOMSdata.R
\name{info_epiPOMSdata}
\alias{info_epiPOMSdata}
\title{Modify and extract information from partially observed multi-strain epidemic
data.}
\usage{
info_epiPOMSdata(epidata)
}
\arguments{
\item{epidata}{An object of class \sQuote{epiPOMSdata}, produced from the
\code{\link{as_epiPOMSdata}} function.}
}
\value{
A list is returned with the following components:
\describe{
\item{ntypes}{Total number of different strains in the study,
\eqn{n_g}.}
\item{obserdata}{A list of the observed test results similar to
\code{epidata}, but with "-" replaced by 0 and "+" by \eqn{(n_g + 1)}.}
\item{ngroups}{Total number of groups in the study, \eqn{P}.}
\item{ninds}{A vector of length \eqn{P}, with each
element \eqn{p} corresponding to the number of individuals in
group \eqn{p}.}
\item{tmax}{A vector of length \eqn{P}, with each
element \eqn{p} corresponding to the time point that the last sample
was collected from the individuals in group \eqn{p}.}
\item{indNA}{A vector of length \eqn{P}, with each element \eqn{p}
corresponding to the index of the individual that withdrawn from
the study before its completion in group \eqn{p}, if any.
Otherwise a value of 0 is given.}
\item{tmaxNA}{A list containing \eqn{P} vectors of length equal to
\code{ninds}\eqn{[p]}, for each \eqn{p = 1, 2, \ldots, P}. Each
element of the vector contains the time point that the last sample
was collected from the corresponding individual in the group, i.e.
if an individual withdrawn from the study before its completion then
this value is less than the time of the last observation in the group.}
}
}
\description{
This function modifies and extracts general information about an object
created by the \code{\link{as_epiPOMSdata}} function.  The output of this
function provides the right format and information required to be used in
other functions in the package.
}
\details{
Only works for objects of class \sQuote{epiPOMSdata}. This function is
used in conjunction with \code{\link{epiPOMS_mcmc}},
\code{\link{plot.epiPOMSmcmc}} and \code{\link{mcmc_initpars}} functions.

Note that, in the event where an individual withdrawn from the study, the
time of the individual's drop-out is recorded as its last non-missing
group pre-specified observation time.
}
\examples{
# Load E. coli O157:H7 data
data(Ecoliepidata)
# Load your own dataframe using myData <- read.csv('myDataFile.csv').

# Format the data
epidata <- as_epiPOMSdata(Ecoliepidata)

# Modify and extract information
infoepidata <- info_epiPOMSdata(epidata)
names(infoepidata)

}
\seealso{
\code{\link{as_epiPOMSdata}} for producing an object of class
\sQuote{epiPOMSdata}, \code{\link{epiPOMS_mcmc}} for performing
inference in partially observed multi-strain epidemic models and
\code{\link{plot.epiPOMSmcmc}} for plotting the outputs of an
\sQuote{epiPOMSmcmc} object.
}
\author{
Panayiota Touloupou
}
