% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_initpars.R
\name{mcmc_initpars}
\alias{mcmc_initpars}
\title{Set starting values for epiPOMS_mcmc MCMC algorithm.}
\usage{
mcmc_initpars(inittype = "prior", priorcontrol, initpars, epidata, smallgroups)
}
\arguments{
\item{inittype}{A character string determining which of the two options to compute the
initial values for the model parameters should be used. Options are
"prior" (the default) and "manual". Setting \code{inittype = } \code{"prior"}
initial values are drawn from their prior distribution; \code{inittype = }
\code{"manual"} initial values are provided by the user. See details, below,
for more information.}

\item{priorcontrol}{A list containing the hyperparameters for the prior
distributions for the MCMC algorithm, obtained using function
\code{\link{prior_control}}. Only used if \code{inittype = } \code{"prior"}.}

\item{initpars}{A list containing the starting values of the model parameters to be
used to initialise the MCMC algorithm, if \code{inittype = } \code{"manual"}.
The names of the components must be "alpha", "beta", "mu", "delta", "gamma",
"nu" and "theta", where:
\describe{
\item{alpha}{is a vector of non-negative real numbers, one for each strain,
containing the initial values of the \eqn{\alpha} parameters, i.e. the
strain-specific external colonisation rates. This is a vector of \eqn{n_g}
values, where \eqn{n_g} is the number of different strains in the study.
The first element corresponds to the first strain, the second element to
the second strain and so on.}
\item{beta}{is a vector of non-negative real numbers, one for each strain,
containing the initial values of the \eqn{\beta} parameters, i.e. the
strain-specific within-group colonisation rates. This is a vector of
\eqn{n_g} values. The first element corresponds to the first strain, the
second element to the second strain and so on.}
\item{mu}{is a vector of non-negative real numbers, one for each strain,
containing the initial values of the \eqn{\mu} parameters, i.e. the
strain-specific clearance rates. This is a vector of \eqn{n_g} values.
The first element corresponds to the first strain, the second element to
the second strain and so on.}
\item{delta}{is a non-negative real number corresponding to the initial value
of the \eqn{\delta} parameter, i.e. the relative colonisation rate in a
carrier versus non-carrier individual.}
\item{gamma}{is a non-negative real number corresponding to the initial value
of the \eqn{\gamma} parameter, i.e. the relative colonisation rate in smaller
versus bigger groups in terms of area (in square meters). If
\code{smallgroups = NULL}, i.e there is no difference between groups, set
\code{gamma = NULL}.}
\item{nu}{is a vector of numbers between 0 and 1, containing the initial values
of the \eqn{\nu} parameters, i.e. the probabilities of carriage at the
beginning of the study. This is a vector of \eqn{(n_g + 1)} values. The first
element corresponds to the non-carriage state. The second and subsequent
elements correspond to the carriage of one of the \eqn{n_g} strains. These
probabilities should sum up to one.}
\item{theta}{is a vector of 5 values, containing the initial values for the
following parameters in order: \eqn{\theta_1}, \eqn{\theta_2}, \eqn{\theta_C},
\eqn{\theta_S} and \eqn{\theta_P}, where \eqn{\theta_1} and \eqn{\theta_2}
denote the test sensitivities. Given that a test is positive, \eqn{\theta_C}
denotes the probability of correctly identifying a common strain,
\eqn{\theta_S} is the probability of misclassifying a common strain with a
different common strain and \eqn{\theta_P} the probability that a strain of
pooled type is classified as a common strain. All parameters must be between
0 and 1 and \eqn{\theta_C + \theta_S \leq 1}{\theta_C + \theta_S \le 1}.}
}}

\item{epidata}{An object of class \sQuote{epiPOMSdata}, produced from the
\code{\link{as_epiPOMSdata}} function.}

\item{smallgroups}{A vector of positive integers, with length equal to the number
of groups that are smaller in terms of area (in square meters), containing
the indices of the small groups. These values must be between 1 and the total
number of groups in the study. Set \code{smallgroups = NULL} if there is no
difference between groups.}
}
\value{
A list is returned with the following components:
\describe{
\item{init_alpha}{A vector of initial values of the \eqn{\alpha}
parameters, i.e. the strain-specific external colonisation rates.
This is a vector of \eqn{n_g} values, where \eqn{n_g} is the number of
strains in the study. The first element corresponds to the first strain,
the second element to the second strain and so on.}
\item{init_beta}{A vector of initial values of the \eqn{\beta}
parameters, i.e. the strain-specific within-group colonisation rates.
This is a vector of \eqn{n_g} values. The first element corresponds to the
first strain, the second element to the second strain and so on.}
\item{init_mu}{A vector of initial values of the \eqn{\mu}
parameters, i.e. the strain-specific clearance rates.
This is a vector of \eqn{n_g} values. The first element corresponds to the
first strain, the second element to the second strain and so on.}
\item{init_delta}{Initial value of the \eqn{\delta} parameter, i.e.
the relative colonisation rate in a carrier versus non-carrier individual.}
\item{init_gamma}{Initial value of the \eqn{\gamma} parameter, i.e.
the relative colonisation rate in smaller versus bigger groups in terms of
area (in square meters), if \code{smallgroups} \eqn{\neq}{≠} \code{NULL}.
Otherwise, \code{init_gamma = NULL}.}
\item{init_nu}{A vector of initial values of the \eqn{\nu} parameters,
i.e. the probabilities of carriage at the beginning of the study. This is a
vector of \eqn{(n_g + 1)} values. The first element corresponds to the
non-carriage state. The second and subsequent elements correspond to the
carriage of one of the \eqn{n_g} strains.}
\item{init_theta}{A vector of initial values of the observation
parameters, in order, \eqn{\theta_1}, \eqn{\theta_2}, \eqn{\theta_C},
\eqn{\theta_S}, \eqn{\theta_P}, where \eqn{\theta_1} and \eqn{\theta_2}
denote the test sensitivities. Given that a test is positive, \eqn{\theta_C}
denotes the probability of correctly identifying a common strain,
\eqn{\theta_S} is the probability of misclassifying a common strain with a
different common strain and \eqn{\theta_P} the probability that a strain of
pooled type is classified as a common strain.}
\item{init_hidprocess}{A list containing \eqn{P} matrices (one for each
group), with the initial hidden carriage process, where \eqn{P} is the total
number of groups in the study. For more information, see
\code{\link{hidstate_sim}}.}
}
}
\description{
Sets starting values used to initialise the Markov chain Monte Carlo (MCMC)
algorithm by \code{\link{epiPOMS_mcmc}} function.
}
\details{
Auxiliary function that can be used to generate initial values for the MCMC
algorithm used by \code{\link{epiPOMS_mcmc}} for performing Bayesian inference
in partially observed multi-strain epidemic models. The output of this function
can be used as an argument of \code{\link{mcmc_control}}.

The argument \code{inittype} has two options. When \code{inittype = }
\code{"prior"} the initial values are generated by sampling each model
parameter from its prior distribution. In this case, the argument
\code{priorcontrol} has to be specified. When  \code{inittype = }
\code{"manual"} the initial values are provided by the user, through
argument \code{initpars}, which has to be specified. The default is to sample
the initial values directly from the prior distribution of the model
parameters. However, in some cases (when using very vague or even improper
flat priors), this is not a good idea since it could generate disparate
initial values that could make the chain take an excessive time to achieve
convergence. Therefore, as a general guideline, it is a safe procedure to give
the initial values of the parameters, at least for those having vague or
improper distributions.
}
\seealso{
\code{\link{epiPOMS_mcmc}} for performing inference in partially observed
multi-strain epidemic models, \code{\link{prior_control}} for specifying the
hyperparameters of the prior distributions and \code{\link{mcmc_control}}
for setting the parameter values that control the MCMC algorithm.
}
\author{
Panayiota Touloupou
}
