% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obserdata_sim.R
\name{obserdata_sim}
\alias{obserdata_sim}
\title{Simulate partially observed multi-strain epidemic data.}
\usage{
obserdata_sim(ntypes, ngroups, ninds, tmax, hidpars, obserpars, smallgroups,
  obsertimes, indNA, tNA, npostyped)
}
\arguments{
\item{ntypes}{An integer greater than 1 providing the number of different strains in
the study, \eqn{n_g}.}

\item{ngroups}{A positive integer providing the total number of groups in the study,
\eqn{P}.}

\item{ninds}{A vector of integers greater than 1 with length \eqn{P}, where
element \eqn{p} gives the number of individuals in group \eqn{p}.}

\item{tmax}{A vector of integers greater than 1 with length \eqn{P}, where
element \eqn{p} gives the time point that the last sample was collected from
individuals in group \eqn{p}.}

\item{hidpars}{A list of parameters with which the simulations of the hidden
process are to be performed. The names of the components must be "alpha",
"beta", "mu", "delta", "gamma" and "nu" (see example). Also, all rates must
be specified in days. These components are:
\describe{
\item{alpha}{A vector of non-negative real numbers, one for each strain,
containing the \eqn{\alpha} parameters, i.e. the strain-specific
external colonisation rates. This is a vector of \eqn{n_g} values.
The first element corresponds to the first strain, the second element to
the second strain and so on.}
\item{beta}{A vector of non-negative real numbers, one for each strain,
containing the \eqn{\beta} parameters, i.e. the strain-specific
within-group colonisation rates. This is a vector of \eqn{n_g} values.
The first element corresponds to the first strain, the second element to
the second strain and so on.}
\item{mu}{A vector of non-negative real numbers, one for each strain,
containing the \eqn{\mu} parameters, i.e. the strain-specific clearance
rates. This is a vector of \eqn{n_g} values.
The first element corresponds to the first strain, the second element to
the second strain and so on.}
\item{delta}{A non-negative real number corresponding to the \eqn{\delta}
parameter, i.e. the relative colonisation rate in a carrier versus
non-carrier individual.}
\item{gamma}{A non-negative real number corresponding to the \eqn{\gamma}
parameter, i.e. the relative colonisation rate in smaller versus bigger
groups in terms of area (in square meters). If
\code{smallgroups = NULL}, i.e there is no difference between groups, set
\code{gamma = NULL}.}
\item{nu}{A vector of numbers between 0 and 1, containing the \eqn{\nu}
parameters, i.e. the probabilities of carriage at the beginning of the study.
This is a vector of \eqn{(n_g + 1)} values. The first element corresponds to
the non-carriage state. The second and subsequent elements correspond to the
carriage of one of the \eqn{n_g} strains. These probabilities should sum up to
one.}
}}

\item{obserpars}{A vector of 5 values, with which the simulations of the observed
process are to be performed, containing the following parameters in order:
\eqn{\theta_1}, \eqn{\theta_2}, \eqn{\theta_C}, \eqn{\theta_S} and
\eqn{\theta_P}, where \eqn{\theta_1} and \eqn{\theta_2} denote the test
sensitivities. Given that a test is positive, \eqn{\theta_C} denotes
the probability of correctly identifying a common strain, \eqn{\theta_S}
is the probability of misclassifying a common strain with a different common
strain and \eqn{\theta_P} the probability that a strain of pooled type is
classified as a common strain. All parameters must be between 0 and 1 and
\eqn{\theta_C + \theta_S \leq 1}{\theta_C + \theta_S \le 1}. See details,
below, for more information about test classification.}

\item{smallgroups}{A vector of positive integers, with length equal to the number
of groups that are smaller in terms of area (in square meters), containing
the indices of the small groups. These values must be between 1 and \eqn{P}.
Set \code{smallgroups = NULL} if there is no difference between groups.}

\item{obsertimes}{A list containing \eqn{P} vectors of positive integers
giving the observation/sampling times, i.e. the times that individuals
were tested in each group \eqn{p = 1, 2, \dots, P}, with entries ranging
from one to \code{tmax}\eqn{[p]}. Each time vector must be in ascending
order, have length greater than 1 and first element equal to 1 (corresponding
to the time that the first sample was collected from the individuals in that
group). Moreover, the last element of each vector \eqn{p} must be equal to
\code{tmax}\eqn{[p]} (see example).}

\item{indNA}{A vector of non-negative integers with length \eqn{P}, where element
\eqn{p} gives the index of the individual that withdrawn from the study
before its completion in group \eqn{p}, if any. Otherwise a value
of 0 should be given. Note that at most one individual from each group can
drop-out before the study period ends.}

\item{tNA}{A vector of positive integers with length \eqn{P}, where element \eqn{p}
provides the individual's drop-out time point in group \eqn{p}, if any,
otherwise the time of the last observation in group \eqn{p} should be given,
\code{tmax}\eqn{[p]}. Note that, if a value is less than the time
of the last observation the corresponding value in \code{indNA} must be
greater than 0, i.e. if \code{tNA}\eqn{[p]} < \code{tmax}\eqn{[p]}
then \code{indNA}\eqn{[p]} > 0.}

\item{npostyped}{A positive integer providing the number of positive samples
(from either test) that will be randomly selected to be strain typed from
each group of the study. When less than \code{npostyped} positive samples
obtained within a group, all the available samples are chosen to be typed.}
}
\value{
An object of class \sQuote{epiPOMSdata} is returned containing the simulated
observed test results of all individuals. For a precise description,
see \code{\link{as_epiPOMSdata}}.
}
\description{
This function allows the user to generate simulations of the measurement
process (the observed test results) of a population of individuals which is
partitioned into groups. This function returns an object of class
\sQuote{epiPOMSdata}.
}
\details{
Our approach to analyse infection dynamics in longitudinal studies where
the process is only partially observed, in the sense that the data are
subject to potential testing error due to poor sensitivity of the
diagnostic and strain procedure used, involves assuming that the
classifications at an observation time are imperfect measures of an underlying
true (hidden) epidemic process. Moreover, our approach accounts for missing
observations, for example, sparse sampling intervals or individual dropouts.

As noted above, the underlying carriage process is not directly observed.
Instead, for each individual we obtain the results of 2 diagnostic tests
that are taken on pre-specified discrete times in each group. Function
\code{\link{obserdata_sim}} simulates the observed process within
each group for each time point over the sampling period. It should be noted
that the interval between two time points is one day. At each sampling time
the observed states are generated conditional on the true disease
states, as obtained by function \code{\link{hidstate_sim}}, according
to a misspecification matrix (one for each test). Once the complete data set
of fully strain typed test results is generated, we randomly choose
\code{npostyped} strain typed observations per group to remain typed and set
the remaining as untyped positives. When less than \code{npostyped} typed
samples occured within a group, all samples remain typed.

To sum up, at an observation time a test result is classified as "-", if
it is negative, "+" if it is positive but not chosen for strain typing,
otherwise, strain number is given ranging from 1 to \eqn{n_g}, where
1 to \eqn{(n_g - 1)} refers to carriage of one of the main strains and
\eqn{n_g} to carriage of the remaining strains (pooled group). \code{NA}
entries corresponds to missing values, either because the individual is
not tested, for example due to sparse sampling intervals, or withdraw from
the study before its completion. A detailed description is found in
\insertCite{Touloupou2020;textual}{epiPOMS}.
}
\examples{
# Simulate partially observed multi-strain epidemic data. Individuals in
# 15 different groups of size varied from 4 to 9, were tested at the same
# time points, every 3 days during a period of 100 days. Five strains
# were identified.
set.seed(1)
ntypes <- 5
ngroups <- 15
ninds <- sample(4:9, size = ngroups, replace = TRUE)
tmax <- rep(100, ngroups)
hidpars <- list(alpha = c(sample(c(0.0015, 0.001), ntypes - 1, replace = TRUE),
     0.002), beta = sample(c(0.005, 0.01), ntypes, replace = TRUE), mu =
     sample(c(0.1, 0.15), ntypes, replace = TRUE), delta = 0.5, gamma = 2,
     nu = c(0.9, rep((1-0.9)/(ntypes+1), (ntypes-1)), 2*(1-0.9)/(ntypes+1)))
obserpars <- c(0.8, 0.5, 0.8, 0.025, 0.05)
smallgroups <- sample(1:ngroups, size = 10)
obsertimes <- lapply(tmax, function(x) seq(1, x, 3))
indNA <- c(rep(0, ngroups-1), 1)
tNA <- c(rep(100, ngroups-1), 70)
npostyped <- 20
epidata <- obserdata_sim(ntypes = ntypes, ngroups = ngroups, ninds = ninds,
     tmax = tmax, hidpars = hidpars, obserpars = obserpars, smallgroups =
     smallgroups, obsertimes = obsertimes, indNA = indNA, tNA = tNA,
     npostyped = npostyped)

# Plot the simulated epidemic data
plot(epidata)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{epiPOMS_mcmc}} for performing inference in partially
observed multi-strain epidemic models.
}
\author{
Panayiota Touloupou, Simon Spencer
}
