% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.epiPOMSdata.R
\name{plot.epiPOMSdata}
\alias{plot.epiPOMSdata}
\title{Plot partially observed multi-strain epidemic data.}
\usage{
\method{plot}{epiPOMSdata}(x, options_gen = list(main = sprintf("Group \%g",
  1:length(x$epidatatest1)), mfrow = c(2, 2), ask = dev.interactive()),
  options_x = list(xlim = lapply(x$epidatatest1, function(x) c(1,
  dim(x)[2])), xlab = "Time", xticks = TRUE), options_y = list(ylim =
  lapply(x$epidatatest1, function(x) c(1, dim(x)[1])), ylab = "Individual",
  yticks = TRUE), options_tests = list(cex = 0.6, col = c("red", "blue")),
  ...)
}
\arguments{
\item{x}{An object of class \sQuote{epiPOMSdata}, produced from the
\code{\link{as_epiPOMSdata}} function.}

\item{options_gen}{A list of general graphical options:
\describe{
\item{main}{A string vector of length \eqn{P}, the total number of groups in
the study, specifying the main title for each plot. Defaults to
\code{sprintf("Group \%g", 1:length(} \code{x$epidatatest1))}.}
\item{mfrow}{A vector of two positive integers of the form \code{c(nr, nc)},
similar to that in \code{par}, corresponding to the number of rows and columns
in the resulting plotting matrix, i.e. subsequent figures will be drawn in an
\code{nr}-by-\code{nc} array on the device. The default
values are 2 for creating a 2\eqn{\times}{x}2 plotting matrix.}
\item{ask}{A logical value similar to that in \code{par}. If
\code{TRUE} (and the \R session is interactive) the user will be prompted
before a new page of plots is started. The default is \code{ask =}
\code{dev.interactive()}.}
}}

\item{options_x}{A list of graphical options for the horizontal axis:
\describe{
\item{xlim}{A list containing \eqn{P} numeric vectors of length 2, giving
the x coordinates ranges for each plot (similar to that in
\code{plot}). Defaults to \code{lapply(x$epidatatest1,}
\code{function(x)} \code{c(1,dim(x)[2]))}, i.e. from 1 to the time of the
last observation in each group.}
\item{xlab}{A label for x-axis. Defaults to "Time".}
\item{xticks}{A logical value; if \code{TRUE}, tick marks on x-axis are
displayed at every observation time point per group. If it is \code{FALSE},
then the default style of the generic \code{plot} function
for x-axis interval calculation is used, where tick marks are equally spaced.
The default is \code{TRUE}.}
}}

\item{options_y}{A list of graphical options for the vertical axis:
\describe{
\item{ylim}{A list containing \eqn{P} numeric vectors of length 2, giving
the y coordinates ranges for each plot (similar to that in
\code{plot}). Defaults to \code{lapply(x$epidatatest1,}
\code{function(x)} \code{c(1,dim(x)[1]))}, i.e. from 1 to the total number of
individuals per group.}
\item{ylab}{A label for y-axis. Defaults to "Individual".}
\item{yticks}{A logical value; if \code{TRUE}, tick marks on y-axis are
displayed at every individual index per group. If it is \code{FALSE}, then
the default style of the generic \code{plot} function for
y-axis interval calculation is used. The default is \code{TRUE}.}
}}

\item{options_tests}{A list of graphical options related to test results:
\describe{
\item{cex}{A positive numerical value, similar to that in \code{plot}, giving
 the amount by which the observed test result labels should be magnified.
 Defaults to 0.6.}
\item{col}{A vector of two colours used for plotting the individual's observed
results of test 1 and test 2, respectively. By default uses red for test 1
and blue for test 2.}
}}

\item{...}{Additional arguments that are passed to the generic
\code{plot} function.}
}
\value{
A series of plots, containing the observed test results for each
individual over the sampling time with separate plots for each group.
The first line for each individual represents the outcome of the first test
and the second line the outcome of the second test.
}
\description{
This function plots multi-strain epidemic longitudinal data consists
of a set of result sequences from two different diagnostic tests among a
population of individuals which is partitioned into groups. The data
should be in the form of the output produced by \code{\link{as_epiPOMSdata}}.
}
\details{
Only works for objects of class \sQuote{epiPOMSdata}. The default plotting
parameter values work well for epidemics up to about 10 individuals per
group over the sampling period of about 100 days. For a larger number of
individuals per group or sampling period, it is recommended to
change the plotting matrix values or use \code{pdf}
and adjust the plotting dimensions.
}
\examples{
# Load E. coli O157:H7 data
data(Ecoliepidata)
# Load your own dataframe using myData <- read.csv('myDataFile.csv').

# Format the data
epidata <- as_epiPOMSdata(Ecoliepidata)

# Plot the data
plot(epidata)

}
\seealso{
\code{\link{as_epiPOMSdata}} for producing an object of class
\sQuote{epiPOMSdata}.
}
\author{
Panayiota Touloupou
}
