% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_control.R
\name{prior_control}
\alias{prior_control}
\title{Set hyperparameters for the prior distributions for epiPOMS_mcmc MCMC
algorithm.}
\usage{
prior_control(ntypes, alphaprior = 1, betaprior = 1, muprior = 1,
  deltaprior = log(2), gammaprior = log(2), nuprior = rep(0.5, ntypes +
  1), theta12prior = c(0.5, 0.5), thetaCSprior = c(4.5, 0.25, 0.25),
  thetaPprior = c(0.5, 0.5))
}
\arguments{
\item{ntypes}{An integer greater than 1 providing the number of different strains in
the study, \eqn{n_g}.}

\item{alphaprior}{A positive real number corresponding to the rate of the Exponential
prior distribution of the \eqn{\alpha} parameters, i.e.
the strain-specific external colonisation rates. Defaults to 1. See details,
below, for more information.}

\item{betaprior}{A positive real number corresponding to the rate of the Exponential
prior distribution of the \eqn{\beta} parameters, i.e.
the strain-specific within-group colonisation rates. Defaults to 1. See
details, below, for more information.}

\item{muprior}{A positive real number corresponding to the rate of the Exponential
prior distribution of the \eqn{\mu} parameters, i.e.
the strain-specific clearance rates. Defaults to 1. See details,
below, for more information.}

\item{deltaprior}{A positive real number corresponding to the rate of the Exponential
prior distribution of the \eqn{\delta} parameter, i.e.
the relative colonisation rate in a carrier versus non-carrier individual.
Defaults to \eqn{\ln(2)}{ln(2)}. See details, below, for more information.}

\item{gammaprior}{A positive real number corresponding to the rate of the Exponential
prior distribution of the \eqn{\gamma} parameter, i.e. the relative
colonisation rate in smaller versus bigger groups in terms of area (in square
meters). If there is no difference between groups, set \code{gammaprior =
NULL}. Defaults to \eqn{\ln(2)}{ln(2)}. See details, below, for more
information.}

\item{nuprior}{A vector of positive real numbers providing the shape parameters of
the joint Dirichlet prior distribution of the \eqn{\nu} parameters, i.e.
the probabilities of carriage at the beginning of the study. This is a vector
of \eqn{n_g + 1} values. The first element corresponds to the non-carriage
state. The second and subsequent elements correspond to the carriage of one of
the \eqn{n_g} strains.  The default value for all \code{nuprior}
elements is 0.5. See details, below, for more information.}

\item{theta12prior}{A vector of two positive real numbers providing the shape
parameters of the Beta prior distribution of the \eqn{\theta_1} and
\eqn{\theta_2} parameters, i.e. the test sensitivities. By default,
\code{theta12prior = c(0.5, 0.5)}. See details, below, for more
information.}

\item{thetaCSprior}{A vector of three positive real numbers providing the shape
parameters of the joint Dirichlet prior distribution of the \eqn{\theta_C} and
\eqn{\theta_S} parameters, i.e. given that a test is found positive
\eqn{\theta_C} denotes the probability of correctly identifying a
common strain \eqn{(1, 2, \ldots, n_g - 1)} and \eqn{\theta_S} is the
probability of misclassifying a common strain with a different common strain.
The default values for the \code{thetaCSprior} elements are (4.5, 0.25, 0.25).
See details, below, for more information.}

\item{thetaPprior}{A vector of two positive real numbers providing the shape
parameters of the Beta prior distribution of the \eqn{\theta_P} parameter,
i.e. the probability that a strain of pooled type \eqn{n_g} is
classified as a common strain given that a test is found positive.
By default, \code{thetaPprior = c(0.5, 0.5)}. See details, below, for more
information.}
}
\value{
A list with components \code{alphaprior}, \code{betaprior}, \code{muprior},
\code{deltaprior}, \code{gammaprior}, \code{nuprior}, \code{theta12prior},
\code{thetaCSprior} and \code{thetaPprior}. This can be used as an argument
of functions \code{\link{epiPOMS_mcmc}} and \code{\link{mcmc_initpars}}.
}
\description{
Sets the hyperparameters for the prior distributions to be used in the
\code{\link{epiPOMS_mcmc}} Markov chain Monte Carlo (MCMC) algorithm to
perform Bayesian inference in partially observed multi-strain epidemic
models.
}
\details{
Auxiliary function that can be used to set hyperparameter values for the
prior distributions for the MCMC algorithm used by \code{\link{epiPOMS_mcmc}}
for performing Bayesian inference in partially observed multi-strain epidemic
models.

For the strain-specific external colonisation rates (\eqn{\alpha}), the
within-pen colonisation rates (\eqn{\beta}) and the clearance rates
(\eqn{\mu}), we assign univariate Exponential priors each with default rate
equal to 1. The priors for \eqn{\delta} and \eqn{\gamma} are also assumed to
be Exponential with default rate equal to \eqn{\ln(2)}{ln(2)}, reflecting
equal prior probabilities for these parameters to be less or more than
one. By default parameter \eqn{\delta} is included in the model. However,
if there is no difference between groups, we allow for the possibility to
exclude this parameter for the model by setting \code{gammaprior = NULL}.
For the misspecification parameters \eqn{\theta_1}, \eqn{\theta_2} and
\eqn{\theta_P} we assume Beta prior distributions, with default
of Beta(0.5, 0.5) which is the Jeffreys' prior
\insertCite{Jeffreys1961}{epiPOMS}. For the remaining observation
parameters we assume a joint Dirichlet prior distribution, with default the
\eqn{(\theta_C,} \eqn{\theta_S,} \eqn{1 - \theta_C - \theta_S)}  \eqn{\sim}{~}
Dirichlet(4.5, 0.25, 0.25) minimally informative prior
\insertCite{Kelly2011}{epiPOMS}. Finally, for the
probabilities of carriage at the beginning of the study (\eqn{\nu}) we use
a joint Dirichlet prior with a default the Jeffreys' Dirichlet
distribution with all \eqn{n_g + 1} parameters set to 0.5. For the full
details see \insertCite{Touloupou2020;textual}{epiPOMS}.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{epiPOMS_mcmc}} for performing inference in partially observed
multi-strain epidemic models and \code{\link{mcmc_initpars}} for producing
the initial values for the MCMC algorithm.
}
\author{
Panayiota Touloupou
}
