% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.epiPOMSmcmc.R
\name{summary.epiPOMSmcmc}
\alias{summary.epiPOMSmcmc}
\title{Summarize posterior samples from \sQuote{epiPOMSmcmc} object.}
\usage{
\method{summary}{epiPOMSmcmc}(object, HPDprob = 0.9, quantprobs = c(0.05,
  0.25, 0.5, 0.75, 0.95), ask = interactive(), ...)
}
\arguments{
\item{object}{An object of class \sQuote{epiPOMSmcmc}, produced from the
\code{\link{epiPOMS_mcmc}} function.}

\item{HPDprob}{A numeric scalar in the interval (0,1), similar to that in
\code{HPDinterval} from the \code{coda} package, giving the target probability
content of the highest posterior density intervals. Defaults to 0.9.}

\item{quantprobs}{A numeric vector of probabilities with values in [0,1], similar to
that in \code{quantile} from the \code{stats} package, producing sample
quantiles corresponding to the given probabilities. The default is
\code{c(0.05, 0.25, 0.5, 0.75, 0.95)}.}

\item{ask}{A logical value; If \code{TRUE} (and the \R session is interactive) the
user will be prompted before a new output is printed. The default is
\code{ask = interactive()}.}

\item{...}{Additional arguments, such as \code{digits}, that are passed to the
generic \code{summary} function.}
}
\value{
Returns and prints summaries of posterior samples generated by the
\code{\link{epiPOMS_mcmc}} inference routine, including the highest posterior
density intervals (\code{HPD}), quantiles (\code{quantiles}), effective sample
size (\code{ESS}) for each parameter, and the mean (\code{mean_vec}) and
covariance matrix (\code{cov_mat}) of all parameters.
}
\description{
The summary method of \sQuote{epiPOMSmcmc} objects (i.e. the output of the
\code{\link{epiPOMS_mcmc}} function) can be used to display and return four
sets of summary information. The summary includes highest posterior density
(HPD) intervals, quantiles and effective sample size (ESS) of the sample
distribution for each model parameter, as well as, the mean vector and the
covariance matrix of the posterior distribution of the parameters.
}
\details{
The highest posterior density intervals and effective sample size for each
model parameter are computed using the \code{HPDinterval} and
\code{effectiveSize} function, respectively, from the \code{coda} package.
}
\seealso{
\code{\link{epiPOMS_mcmc}} for generating posterior samples of the
model parameters and \code{\link{plot.epiPOMSmcmc}} for
plotting the outputs of an \sQuote{epiPOMSmcmc} object.
}
\author{
Panayiota Touloupou
}
